# openapi.json
{"openapi":"3.1.0","info":{"title":"GlowGuard - AI Skin Care Platform","description":"AI-powered skin disease prediction and skincare guidance platform","version":"1.0.0"},"paths":{"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login user","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Logout user - token invalidation handled on frontend","operationId":"logout_api_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/products/":{"get":{"tags":["Products"],"summary":"Get All Products","description":"Get all products","operationId":"get_all_products_api_products__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/products/by-category/{category}":{"get":{"tags":["Products"],"summary":"Get Products By Category","description":"Get products by category","operationId":"get_products_by_category_api_products_by_category__category__get","parameters":[{"name":"category","in":"path","required":true,"schema":{"type":"string","title":"Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/recommended/{disease}":{"get":{"tags":["Products"],"summary":"Get Recommended Products","description":"Get products recommended for a specific disease","operationId":"get_recommended_products_api_products_recommended__disease__get","parameters":[{"name":"disease","in":"path","required":true,"schema":{"type":"string","title":"Disease"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/{product_id}":{"get":{"tags":["Products"],"summary":"Get Product","description":"Get specific product","operationId":"get_product_api_products__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/recommendations/by-prediction/{prediction_id}":{"get":{"tags":["Recommendations"],"summary":"Get Recommendations By Prediction","description":"Get recommendations for a specific prediction","operationId":"get_recommendations_by_prediction_api_recommendations_by_prediction__prediction_id__get","parameters":[{"name":"prediction_id","in":"path","required":true,"schema":{"type":"integer","title":"Prediction Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/recommendations/by-category/{category}":{"get":{"tags":["Recommendations"],"summary":"Get Recommendations By Category","description":"Get all recommendations by category","operationId":"get_recommendations_by_category_api_recommendations_by_category__category__get","parameters":[{"name":"category","in":"path","required":true,"schema":{"type":"string","title":"Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/profile":{"get":{"tags":["Users"],"summary":"Get Profile","description":"Get current user profile","operationId":"get_profile_api_users_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]},"put":{"tags":["Users"],"summary":"Update Profile","description":"Update user profile","operationId":"update_profile_api_users_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/users/account":{"delete":{"tags":["Users"],"summary":"Delete Account","description":"Delete user account","operationId":"delete_account_api_users_account_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","token_type","user"],"title":"TokenResponse"},"UserCreate":{"properties":{"email":{"type":"string","pattern":"^[\\w\\.-]+@[\\w\\.-]+\\.\\w+$","title":"Email"},"username":{"type":"string","title":"Username"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"skin_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skin Type"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","username","password"],"title":"UserCreate"},"UserResponse":{"properties":{"email":{"type":"string","pattern":"^[\\w\\.-]+@[\\w\\.-]+\\.\\w+$","title":"Email"},"username":{"type":"string","title":"Username"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"skin_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skin Type"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["email","username","id","created_at","is_active"],"title":"UserResponse"},"UserUpdate":{"properties":{"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"skin_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skin Type"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}